Removing refactor in favor of getting it to work first
